every part now handles the situation where one or both inputs are empty

+++ panoptica/_functionals.py
@@ -91,4 +91,4 @@ def _connected_components(
     else:
         raise NotImplementedError(cca_backend)
 
-    return cc_arr, n_instances
+    return cc_arr.astype(array.dtype), n_instances
